power: Rise max cpu frequency to 2.5Ghz
In flounder we have a cpu cap that limits the cpu clock from going over 2.3ghz.
Having the powerhal set it that low does not work for us because if the user
sets the cap and cpu freq higher than what the powerhal had set it will cause
the cpu speed to reset back to 2.3ghz if the screen goes off.
We don't want this, we want to keep what the user had set.
To fix this we have to set the freq to 2.5ghz but the cpu cap
will still keep the cpu freq stock/2.3ghz.

Change-Id: I4bf4950f04ad4b1fced5a6d16e7cc78a22e7681a
Signed-off-by: USBhost <josiahspore@gmail.com>
Signed-off-by: André Pinela <sheffzor@gmail.com>