Update patch set 1
Patch Set 1:

> Is this shipped by google?

The thing is we only got grid recents on Android 7.1.2. Our Nexus 9 never was updated to .2 so we wouldn't know. The Pixel C got it when it updated to .2 and its a tablet so as well is our Nexus 9. But to answer your question no it was not shipped by google.

Patch-set: 1
Reviewer: Josiah Morauske <1000224@d5e41ece-075f-43fb-b5a6-48951f3d7b1d>
Label: Code-Review=0